Another round of refactoring
