+++ Nbuild/Nbuild.csproj
@@ -45,7 +45,9 @@
 	
 	<ItemGroup>
 	  <PackageReference Include="CommandLine.Net" Version="2.3.0" />
-	  <PackageReference Include="ntools-launcher" Version="1.0.0.5" />
 	</ItemGroup>
+ <ItemGroup>
+    <ProjectReference Include="..\launcher\Launcher.csproj" />
+  </ItemGroup>
 
 </Project>
