+++ launcher/launcher.csproj
@@ -6,7 +6,8 @@
     <AssemblyVersion></AssemblyVersion>
     <Authors>naz-hage</Authors>
     <Company>naz-hage</Company>
-    <Product>launcher</Product>
+    <Product>Ntools</Product>
+    <Description>A wrapper for Process .NET class</Description>
     <Copyright>2020-2024</Copyright>
   </PropertyGroup>
 
@@ -26,18 +27,6 @@
   <PropertyGroup>
     <GenerateAssemblyInfo>true</GenerateAssemblyInfo>
     <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
-    <GeneratePackageOnBuild>False</GeneratePackageOnBuild>
-    <PackageId>ntools-$(AssemblyName)</PackageId>
-    <Title>A wrapper for Process .NET class</Title>
-    <Description>This package simplifies calling another executable. </Description>
-    <PackageProjectUrl>https://github.com/naz-hage/ntools/tree/main/launcher</PackageProjectUrl>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
-    <RepositoryUrl>https://github.com/naz-hage/ntools/tree/main/launcher</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
-    <PackageReleaseNotes>First release</PackageReleaseNotes>
-    <PackageOutputPath>$(MSBuildProjectDirectory)\..\nupkgs</PackageOutputPath>
-    <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
-    <PackageRequireLicenseAcceptance>True</PackageRequireLicenseAcceptance>
   </PropertyGroup>
 
   <ItemGroup>
